requirements.txt
pytorch
matplotlib
wandb
pyyaml
gym
mujoco
dm-control